Zero torque after timeout
Add a timeout to the ZMQ command loop to act as a deadman's switch.

When the last received command is more than a defined time interval
in the past, it zeros the commanded joint torque. This prevents a
communication error or stopped external controller from persisting
a potentially dangerous torque value.

Timeout is currently defined with chrono types as 500 milliseconds