Torque timeout (#3)
* Zero torque after timeout

Add a timeout to the ZMQ command loop to act as a deadman's switch.

When the last received command is more than a defined time interval
in the past, it zeros the commanded joint torque. This prevents a
communication error or stopped external controller from persisting
a potentially dangerous torque value.

Timeout is currently defined with chrono types as 500 milliseconds

* Flush command port on stop

* Poll the command port when the robot stops (e.g. when
run_joint_torques_controller fails) to flush the last
message. Also set command torque to zero in this case.

If the robot is stopped (with the switch or with a safety limit)
before the external controller is stopped, it is likely for a
high torque command message to persist on the port. The
next time the robot starts, it will poll the port, see the last
message and update the command (and last command time).
Particularly in cases where a high torque exceeded the safety
limit of the robot, this means that reactivating robot would
cause a sudden jerky motion.

Now, the system is robust to both types of stops: either the
external controller stops sending, in which case the running
robot zeros the torque command after a timeout, or the
robot stops before the controller, in which case the port
is flushed and the command is zeroed prior to restart.