Add ControlType to protocol and interface
* Add ControlType enum to communication protocol
header, with NONE = 0 and types for joint or
cartesian position, velocity or force control.
Not all control types are implemented, but
the types are defined already and reserved
for future use.

* Add helper data() access functions to
Vec3D, Quaternion, EEPose, EETwist structs
to return a std array. Mark conversion
helpers as deprecated.

* Add more command members to the
interface to capture all potential command
types.

* Add state logic to interface to switch
controller based on control type.

* Add run_state_publisher default "controller"
that publishes state data and runs when the
control type is NONE or unimplemented.

* Add run_cartesian_velocities_controller that
uses the Franka cartesian impedance controller
with a twist command.

* Add project version 1.1 in CMakeList, and define
FRANKA_LWI_PROTO_VERSION 1.1 in header